Adding --sleep and --first_block (not implemented) to chifra init

+++ src/apps/chifra/internal/scrape/options.go
@@ -38,7 +38,7 @@ func (opts *ScrapeOptions) testLog() {
 	logger.TestLog(opts.BlockCnt != 2000, "BlockCnt: ", opts.BlockCnt)
 	logger.TestLog(opts.Pin, "Pin: ", opts.Pin)
 	logger.TestLog(opts.Remote, "Remote: ", opts.Remote)
-	logger.TestLog(opts.Sleep != 14, "Sleep: ", opts.Sleep)
+	logger.TestLog(opts.Sleep != float64(14), "Sleep: ", opts.Sleep)
 	logger.TestLog(opts.StartBlock != 0, "StartBlock: ", opts.StartBlock)
 	opts.Settings.TestLog(opts.Globals.Chain, opts.Globals.TestMode)
 	opts.Globals.TestLog()
